requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.8.2
pydantic-settings==2.4.0
passlib==1.7.4
argon2-cffi==23.1.0
python-jose==3.3.0
email-validator==2.2.0
redis==5.0.8
rq==1.16.2
boto3==1.35.6
cryptography==43.0.1
httpx==0.27.2
structlog==24.4.0
prometheus-client==0.20.0
slowapi==0.1.9
web3>=6.11.0
python-dotenv==1.0.1
PyJWT==2.9.0
pyotp==2.9.0
